+++ packages/admin/client/views/panels.js
@@ -1,17 +1,4 @@
 
-Template.panelPlace_admin.onRendered(function() {
-	var self = this;
-	self.$('.btn-editdel').btsConfirmButton(function(e) {
-			
-		K.Admin.removePlace(self.data.id)
-		
-		K.Map.removeItem(self.data);
-	
-		Router.go('root');
-	});
-});
-
-
 Template.panelUser_admin.onRendered(function() {
 	var self = this;
 	self.$('.btn-editdel').btsConfirmButton(function(e) {