+++ packages/edit/client/Place_edit.js
@@ -2,6 +2,6 @@
 Kepler.Place.include({
 	
 	isEditable: function() {
-		return this.userId ? K.Profile.id === this.userId : false;
+		return !!(this.userId ? (K.Profile.id === this.userId) : false);
 	}
 });