+++ packages/edit/client/router.js
@@ -11,6 +11,7 @@ Router.map(function() {
 		},
 		data: function() {
 			if(!this.ready()) return null;
+			
 			var place = K.placeById( this.params.placeId );
 
 			if(!place) {