+++ packages/import/client/views/panels.js
@@ -13,12 +13,10 @@ Template.item_place_imported.onRendered(function() {
 	var self = this,
 		tmpl = Template.instance();
 
-	self.$('.btn-editdel').btsConfirmButton(function(e) {
+	self.$('.btn-placedel').btsConfirmButton(function(e) {
 		
 		Meteor.call('removePlace', self.data.id, function(err,res) {
-		console.log(self.data.id, err,res)
-			K.Map.removeItem(self.data);
-		
+			K.Map.removeItem(self.data);		
 			//Router.go('root');
 		});			
 	});