+++ packages/photos/client/views/items.js
@@ -1,9 +1,0 @@
-
-
-
-Template.item_photo_edit.helpers({
-	url: function() {
-		//var tmpl = Template.instance();
-		return '';
-	}
-});