+++ packages/photos/server/places.js
@@ -2,15 +2,57 @@
 
 Meteor.methods({
 	
-	insertPlacePhotos: function(fileObj, sets) {
+	updatePlacePhoto: function(fileObj, placeId) {
 
-		var url = Meteor.call('resizePhoto', fileObj, sets);
+		if(!this.userId) return null;
 
-		var exif = Meteor.call('exifPhoto', fileObj, sets);
+		var sets = K.settings.upload.targets['photos_places'];
 
-		return {
-			url: url,
-			loc: exif.loc
-		};
+		var placeData = Places.findOne(placeId);
+
+		if(placeData.userId === this.userId || (K.Admin && K.Admin.isMe())) {
+				
+			//var url = Meteor.call('resizePhoto', fileObj, sets);
+			//var exif = Meteor.call('exifPhoto', fileObj, sets);
+			
+			var url = Meteor.call('uploadPhoto', fileObj, sets);
+
+			Places.update(placeId, {
+				$set: {
+					photo: url
+				}
+			});
+
+			console.log('Photos: updatePlacePhoto', placeId)
+
+			return url;
+		}
+	},
+
+	removePlacePhoto: function(placeId) {
+
+		if(!this.userId) return null;
+
+		var sets = K.settings.upload.targets['photos_places'];
+
+		var placeData = Places.findOne(placeId);
+
+		if( (placeData && placeData.userId === this.userId) || 
+			(K.Admin && K.Admin.isMe()) ) {
+		
+			//var url = Meteor.call('resizePhoto', fileObj, sets);
+			//var exif = Meteor.call('exifPhoto', fileObj, sets);
+			//var url = Meteor.call('deletePhoto', placeData.photo, sets);
+
+			Places.update(placeId, {
+				$set: {
+					photo: ''
+				}
+			});
+
+			//TODO delete photo file
+
+			console.log('Photos: removePlacePhoto', placeId);
+		}
 	}
 });